🟢 feat: Operand blocks can now be used as parameters!
Any operand block works as it was planned. They don't connect to other blocks on the bottom, instead connecting to any parameter with the Block type. The order the editor checks each block if its being dragged has been reversed, solving a few other problems and making the editor feel and work better.